Remove statusbar button
Statusbar does not exists anymore since Firefox 4, so we remove it. The
toolbar button can be added to the add-on bar, so it won't be an issue.

Also, if the newli added toolbar button should work right after the user
moves the mouse over the icon.
That's a kind of hack, because I don't know about any event which can
tell me if a button has been added to the toolbar. So, on mouse over we
will add the listeners to the button and we will run a check.