Show unread count at icon and in notifier
Getting and displaying unread count works now but it's really not clear.
It will need lot of cleanup and documentation:
Generating the list what could be displayed in the notification should
be much more simple than this.
Also documentation does not exists, it should be written, because even I
don't know why is there some stuff (like friend list).